Update create_dataset.py
Added in a way to programmatically create metadata given a "metadata_index" where the user has to provide a dictionary where {key=coulmn_name : value=desired_metadata_type} -- the docstring explains how to set this up. This protocol also checks to make sure this metadata doesn't exist. This is an optional argument and if you don't provide it, it will create data rows without metadata -- but it will still create / check for the "lb_partner_source" metadata field and will add "Databricks" to it.